requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-enum==2.3.0
django-environ==0.12.0
django-extensions==4.1
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.7
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
